Compacted the favorites list and implemented cookie store
The list now renders top five emojis, which is equal to one row of emoji
using the current style. Forget link was removed and instead a global
erase link is used. Cookies are now used as an alternative storage. Copy
count were reworked to be relative to the list item so they don't share
the save <div> (and the copy count text is not copied with the emoji
code) but also don't break the line.